feat(api): rename file and export for admin-member repository
